requirements.txt
openai
gpt_index
PyPDF2
gradio
atlassian-python-api
pdfkit